move tests to root level & reorganize
