+++ tests/calculations/test_gaussian/test_default.in
