+++ tests/parsers/fixtures/base/nan_inf/aiida.out
