+++ tests/parsers/test_base/test_nan_inf.yml
